Changed parameters of prototypes from int to int8_t

+++ BSP/Inc/BSP_Timer.h
@@ -3,5 +3,5 @@
 
 void (*function1);    //function pointers
 void (*function2);  
-void BSP_Update_Time(int row);
-void BSP_Timer_Init(int reload, void *func1, void *func2);
+void BSP_Update_Time(int8_t row);
+void BSP_Timer_Init(int8_t reload, void *func1, void *func2);
